General refactoring of the TileMap-system
TileSize in GameData is now private and uses getter
Renamed isCoordinateInObstacle --> isCoordinateTraversable in World
Changed TileManager to read number of rows and columns to be less dependent.
Pathfinding now uses the isCoordinateTraversable from World instead of its own local function for better consolidation. This does lead to higher coupling, so not so good.